managed.go
